pyproject.toml
[tool.poetry]
name = "tgbot"
version = "0.1.0"
description = ""
authors = ["maxnemoy <maxnemoy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {extras = ["callback-data"], version = "^20.0a6"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
